feat: Add Gantt, List, Monthly, Timeline, and Weekly views for employee scheduling
- Implemented GanttView component for visualizing employee schedules over a date range.
- Created ListView component to display schedules in a list format with edit and delete options.
- Developed MonthlyView component to show schedules in a calendar format.
- Added TimelineView component to present schedules in a timeline layout with detailed employee information.
- Introduced WeeklyView component to display schedules in a weekly table format with attendance statuses.
- Enhanced filtering capabilities for employee and department selection across all views.
- Added utility functions for cookie management in the client.
- Refactored API service imports for better structure and maintainability.
