+++ src/app/manager/lich-lam-viec/views/index.ts
@@ -1,8 +1,8 @@
-export { default as DailyView } from "./DailyView";
-export { default as WeeklyView } from "./WeeklyView";
-export { default as MonthlyView } from "./MonthlyView";
-export { default as ListView } from "./ListView";
-export { default as CardView } from "./CardView";
-export { default as GanttView } from "./GanttView";
-export { default as TimelineView } from "./TimelineView";
-export { default as DetailView } from "./DetailView";
+export { default as DailyView } from "../../../../components/ViewComponent/DailyView";
+export { default as WeeklyView } from "../../../../components/ViewComponent/WeeklyView";
+export { default as MonthlyView } from "../../../../components/ViewComponent/MonthlyView";
+export { default as ListView } from "../../../../components/ViewComponent/ListView";
+export { default as CardView } from "../../../../components/ViewComponent/CardView";
+export { default as GanttView } from "../../../../components/ViewComponent/GanttView";
+export { default as TimelineView } from "../../../../components/ViewComponent/TimelineView";
+export { default as DetailView } from "../../../../components/ViewComponent/DetailView";
