+++ src/components/ViewComponent/CardView.tsx
