+++ src/components/ViewComponent/DailyView.tsx
